+++ Stucco/template/module/Public/Get-HelloWorld.ps1
@@ -10,12 +10,12 @@ Function Get-HelloWorld {
 
         Runs the command
     #>
+    [OutputType([string])]
     [CmdletBinding()]
     param (
         # Parameter description can go here or above in format: .PARAMETER  <Parameter-Name>
         [Parameter()]
-        [string]
-        $Value = 'GetHelloWorld'
+        [string]$Value = 'GetHelloWorld'
     )
 
     $Value